blame-tree: use exact path name in pathspec
If there is a glob character in a directory name, then that can result
in the path being parsed as a wildcard and thus not valid for
blame-tree. Since wildcards are rejected by blame-tree, we should
instead treat all input as an exact path.

Since we are parsing arguments through the setup_revisions() method, we
need a way to specify to use the PATHSPEC_LITERAL_PATH flag in
parse_pathspec(). Instead of modifying the revisions API and risking
conflicts with upstream changes, use setenv() with the
GIT_LITERAL_PATHSPECS_ENVIRONMENT to use global state to our benefit.

Test that we can use a strange glob name during 'git blame-tree'
directly, but also be sure to use it with the --cache option so that a
follow-up 'git blame-tree --update-cache' command will pick up that path
and call a subcommand correctly. Before this change, the --update-cache
command would return success, but with a warning message:

  fatal: max-depth cannot be used with wildcard pathspecs

This "fatal" message is actually from the 'git blame-tree --cache'
subcommands run by the --update-cache command.