blame-tree: duplicate path from blame-tree cache
The blame-tree cache file format includes a path length parameter, which
shows how to limit the path when parsing the file. The metadata chunk is
also padded to multiples of four, when necessary, with trailing null
bytes. Thus, using the memory map to read the path works _most_ of the
time. However, it doesn't work for paths that perfectly align with four
bytes (using 8 in the test changes in the previous commit is enough to
show this).

By using xstrndup() we can correct this issue.

Without this fix, there is potential that each round of writing
blame-tree caches would add a new, worthless .btc file, since it
involves a longer path. However, this was stopped in production because
of the glob characters that would occasionally find themselves in the
path: these would create a blocker refusing to write the new file,
stopping the otherwise unstoppable chain of files.

While a previous change unlocks the caching of paths with arbitrary
characters, this chaining was also fixed by the change that would delete
.btc files for paths that do not exist at the cached commit.