blame-tree: refuse to --cache without --max-depth
The blame-tree cache format includes a value for the maximum depth, and
this maximum depth is used for computing the name of the .btc file.
However, the current implementation does not distinguish between
"--max-depth=0" and no set maximum depth.

To avoid these collisions, start refusing to write the cache file when
no --max-depth parameter is given. The only way to determine this with
the current parsing logic is to inspect the 'max_depth_valid' member of
the diff options. This member's documentation hits at this fact that
max_depth has a collision at 0.

This only makes a difference for the root path, but if a user requests a
fully-recursive blame-tree for the root path, then it will hit the
--max-depth=0 cache and not have enough results for the full recursive
path list. We need to prevent _reading_ the cache in these scenarios,
too.

If we want to properly enable a max-depth value for full recursion in
the future, we will want to store a -1 (0xffffffff) in place of the max
depth value in the .btc file. These full recursive blame-trees are rare
so should not be prioritized at the moment.