blame-tree: return int from blame_tree_release
One interesting aspect of 'git blame-tree --cache' is that it writes the
cache file after the blame_tree_run() method has completed collecting
all of the results into the blame_tree struct. The blame_tree_release()
method can write or even unlink a blame-tree cache file, and that means
that I/O errors can occur during this time.

Modify the prototype to return an int that passes error codes from the
writing or unlinking mode, then surface that to the builtin so it can
die() to correctly report a failure.