ref-filter: add ref_filter_clear()
We did not bother to clean up at all in branch/tag, and for-each-ref
only hit a few elements. So this is probably cleaning up leaks, but I
didn't check yet.

Note that the reachable_from and unreachable_from lists should be
cleaned as they are used. So this is just covering any case where we
might bail before running the reachability check.