builtin/repack.c: convert PACK_CRUFT into a boolean
We use the third bit of `pack_everything` to indicate whether or not the
user asked for cruft packs or not.

In a future commit, the decision of whether or not to use cruft packs
will become a tri-state. Prepare for that by storing whether or not to
use cruft packs in a separate variable.

Signed-off-by: Taylor Blau <me@ttaylorr.com>