refs/packed-backend.c: implement skip lists to avoid excluded pattern(s)
When iterating through the `packed-refs` file in order to answer a query
like:

    $ git for-each-ref --exclude=refs/__hidden__

it would be useful to avoid walking over all of the entries in
`refs/__hidden__/*` when possible, since we know that the ref-filter
code is going to throw them away anyways.

In certain circumstances, doing so is possible. The algorithm for doing
so is as follows:

  - For each excluded pattern, find the first record that matches it,
    and the first pattern that *doesn't* match it (i.e. the location
    you'd next want to consider when excluding that pattern).

  - Sort the patterns by their starting location within the
    `packed-refs` file.

  - Construct a skip list of regions by combining adjacent and
    overlapping regions from the previous step.

  - When iterating through the `packed-refs` file, if `iter->pos` is
    ever contained in one of the regions from the previous steps,
    advance `iter->pos` past the end of that region, and continue
    enumeration.

Note that this optimization is only possible when none of the excluded
pattern(s) have special meta-characters in them. To see why this is the
case, consider the exclusion pattern "refs/foo[a]". In general, in order
to find the location of the first record that matches this pattern, we
could only consider up to the first meta-character, "refs/foo". But this
doesn't work, since the excluded region we'd come up with would include
"refs/foobar", even though it is not excluded.

There are a few other gotchas worth considering. First, note that the
skip list is sorted, so once we skip past a region, we can avoid
considering it (or any regions preceding it) again. The member
`skip_pos` is used to track the first next-possible region to jump
through.

Second, note that the exclusion list is best-effort, since we do not
handle loose references, and because of the meta-character issue above.

In repositories with a large number of hidden references, the speed-up
can be significant. Tests here are done with a copy of linux.git with a
reference "refs/pull/N" pointing at every commit, as in:

    $ git rev-list HEAD | awk '{ print "create refs/pull/" NR " " $0 }' |
        git update-ref --stdin
    $ git pack-refs --all

, it is significantly faster to have `for-each-ref` skip over the
excluded references, as opposed to filtering them out after the fact:

    $ hyperfine --warmup=3 \
      'git for-each-ref | grep -vE "^[0-9a-f]{40} [^ ]* refs/pull/"' \
      'git.compile for-each-ref --exclude="refs/pull"'
    Benchmark 1: git for-each-ref | grep -vE "^[0-9a-f]{40} [^ ]* refs/pull/"
      Time (mean ± σ):      2.062 s ±  0.008 s    [User: 1.895 s, System: 0.207 s]
      Range (min … max):    2.047 s …  2.073 s    10 runs

    Benchmark 2: git.compile for-each-ref --exclude="refs/pull"
      Time (mean ± σ):       9.6 ms ±   0.3 ms    [User: 1.4 ms, System: 8.2 ms]
      Range (min … max):     9.0 ms …  10.7 ms    273 runs

    Summary
      'git.compile for-each-ref --exclude="refs/pull"' ran
      215.42 ± 5.97 times faster than 'git for-each-ref | grep -vE "^[0-9a-f]{40} [^ ]* refs/pull/"'

Using the skip list is fairly straightforward (see the changes to
`refs/packed-backend.c::next_record()`), but constructing the list is
not. To ensure that the construction is correct, add a new suite of
tests in t1419 covering various corner cases (overlapping regions,
partially overlapping regions, adjacent regions, etc.).

Co-authored-by: Jeff King <peff@peff.net>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Taylor Blau <me@ttaylorr.com>