builtin/receive-pack.c: avoid enumerating hidden references
Now that `refs_for_each_fullref_in()` has the ability to avoid
enumerating references matching certain pattern(s), use that to avoid
visiting hidden refs when constructing the ref advertisement via
receive-pack.

Note that since this exclusion is best-effort, we still need
`show_ref_cb()` to check whether or not each reference is hidden or not
before including it in the advertisement.

This will provide similar savings as in a previous commit when invoking
`receive-pack` on a repository with a large amount of hidden refs.

Signed-off-by: Taylor Blau <me@ttaylorr.com>