builtin/for-each-ref.c: add `--exclude-hidden` option
For certain callers (e.g., those reimplementing functionality in `git
receive-pack`), it may be convenient to enumerate references which would
be part of the advertisement, respecting `receive.hideRefs`.

Since the advent of the `--exclude` option, callers could do something
like:

    {
      git config --get-all transfer.hideRefs
      git config --get-all receive.hideRefs
    } | sed -e 's//--exclude=&/g' | xargs git for-each-ref

but even this isn't complete, since it doesn't take into account the
negated and stripping features ('!' and '^', respectively).

Instead, add an `--exclude-hidden` option which acts as if all patterns

Signed-off-by: Taylor Blau <me@ttaylorr.com>