contrib/credential/osxkeychain: add `test` target
Add a "test" Makefile recipe for the osxkeychain credential helper to
invoke Git's t0303, which performs tests on the helper itself.

One gotcha when running t0303 with this credential helper is that macOS
requires that you leave the $HOME variable intact so that it can locate
the appropriate user keychain. If not, the credential helper will not be
able to read or write any secrets out of the keychain, rendering the
test useless.

Co-authored-by: Jeff King <peff@peff.net>
Signed-off-by: Jeff King <peff@peff.net>
Signed-off-by: Taylor Blau <me@ttaylorr.com>