blame-tree: reorganize headers
A future change will introduce a use of blame-tree.h in builtin/fsck.c,
but the current unavoidable inclusion of revision.h creates a conflict
in the definition of the SEEN macro.

To avoid this, we could rename the SEEN macro in builtin/fsck, but that
presents opportunities to conflict with upstream merges. Instead, remove
the need for including revision.h in blame-tree.h by moving the full
definition of 'struct blame_tree' to a new blame-tree-internal.h. This
allows the consumers of that struct in blame-tree.c and
builtin/blame-tree.c to use the internal version without exposing too
many internal details to other consumers.

Also clean up some other header includes and a callback definition that
doesn't need to be in the header.