fsck: verify checksums of blame-tree caches
In order to determine if our blame-tree caches have filesystem-level
corruption, it is usually sufficient to verify their checksums. They
have these as a byproduct of using the chunk-format API.

Deeper checks are possible, such as verifying that the files can be
parsed by the chunk-format API or that the results are correct if we
computed them fresh in the moment. This is likely unnecessary unless we
find a bug that could have written these incorrectly (with a valid
checksum).

This is integrated into 'git fsck' in the normal way by adding a bit to
the flags used in the return code. However, I initially attempted
skipping to a larger bit and it turns out that such a change is ignored
when the return code is truncated to a _byte_ and hence causing 'git
fsck' to return with success. This is something to bring up with
upstream, but for now I've created a giant warning comment and messed
with the return code.